+++ net-probe-srv/config/config.py
@@ -1,6 +1,6 @@
 # -*- Mode: Python; python-indent-offset: 4 -*-
 #
-# Time-stamp: <2016-11-01 21:17:41 alex>
+# Time-stamp: <2016-12-03 22:39:38 alex>
 #
 
 """
@@ -23,8 +23,10 @@ class config(object):
 
         """
         self.aHostTable = {}
+        self.aTemplates = {}
         self.outputMethodName = "none"
         self.fileName = "none"
+        self.iTemplateJobsId = 1000
 
         return
 
@@ -42,12 +44,67 @@ class config(object):
             return False
 
     # ----------------------------------------------------------
+    def applyTemplateToHost(self, hostData, tName):
+        """apply the template tName to the hostData
+
+        """
+
+        if not self.aTemplates.__contains__(tName):
+            logging.warning("add a template non existant {} {}".format(hostData['id'], tName))
+            return
+
+        logging.info("applyTemplateToHost {} {}".format(hostData['probename'], tName))
+
+        hjs = None
+
+        t = self.aTemplates[tName]
+        if t.__contains__('jobs'):
+
+            # already some jobs on the probe ? maybe an update only
+            if hostData.__contains__('probename'):
+                probename = hostData['probename']
+                
+                for hkey in self.aHostTable:
+                    h = self.aHostTable[hkey]
+                    
+                    if h.__contains__('probename') and h['probename'] == probename:
+                        if h.__contains__('jobs'):
+                            hjs = h['jobs']
+                            break
+
+            for j in t['jobs']:
+                newJob = j.copy()
+                newJob['id'] = 0
+
+                if hjs != None:
+                    for hj in hjs:
+                        if hj.__contains__('template'):
+                            if hj['template'] == tName:
+                                newJob['id'] = hj['id']
+
+                if newJob['id'] == 0:
+                    newJob['id'] = self.iTemplateJobsId
+                    self.iTemplateJobsId += 1
+
+                newJob['template'] = tName
+
+                if hostData.__contains__('jobs'):
+                    hostData['jobs'].append(newJob)
+                else:
+                    hostData['jobs'] = [ newJob ]
+
+    # ----------------------------------------------------------
     def addHost(self, hostData):
-        """add a host to the database
+        """add a host to the database, if template specified, apply 
+           first the template
 
         """
 
         sId = hostData['id']
+
+        if hostData.__contains__('template'):
+            for t in hostData['template']:
+                self.applyTemplateToHost(hostData, t)
 
         if hostData.__contains__('jobs'):
             jobs = hostData['jobs']
@@ -77,6 +134,25 @@ class config(object):
             self.aHostTable[sId] = {"jobs" : jobs, "probename": probename}
 
     # ----------------------------------------------------------
+    def addTemplate(self, templateData):
+        """add a template from the configuration file to the database
+
+        """
+
+        if not templateData.__contains__('name'):
+            assert False, "no name section in template"
+
+        sName = templateData['name']
+        logging.info(sName)
+
+        if templateData.__contains__('jobs'):
+            jobs = templateData['jobs']
+        else:
+            jobs = {}
+
+        self.aTemplates[sName] = {"jobs" : jobs}
+        
+    # ----------------------------------------------------------
     def loadFile(self, sFile):
         """load host file and update configuraion
 
@@ -94,6 +170,13 @@ class config(object):
         f.close()
 
         conf = json.loads(c)
+
+        # template
+        if conf.__contains__('template'):
+            for t in conf['template']:
+                self.addTemplate(t)
+
+        # group
 
         # probes
         if not conf.__contains__('probe'):