+++ py-net-probe/docker/scratch/run.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+
+echo "raspberry
+raspberry" | passwd pi
+
+set -e
+
+/set_root_pw.sh
+exec /usr/sbin/sshd -D